+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Changelog
+
+
+## [1.0.25] 2021-01.xx
+
+### Added
+- `PLC_Thermostat`: add push support for `get_CurrentHeatingCoolingState`
+- `PLC_Thermostat`: add support for `get_TargetHeatingCoolingState` and `set_TargetHeatingCoolingState` including `mapGetTarget` and `mapSetTarget`
+
+### Fixed
+- `PLC_Thermostat` fixed documentation of `get_CurrentHeatingCoolingState`
+
 ## [1.0.24] - 2021-01-05
 
 ### Added